Added expandAll and collapseAll functionality.
This commit establishes the foundational state management for the
expand state. when the user clicks the expandAll button, the
handleExpandAll function gets invoked and it sets an array with
all the users id to the expand state.

when the user clicks the collapseAll button, the
handleCollapseAll function gets invoked and it sets an empty array
to the expand state.

based on the expand array, the about section will only display for the
users ids that exist in the expand array.

future commits will create a similar functionality that will handle
the expansion of a single user at a time.